refactor(TeamColor): 함수형 인터페이스 필드 삭제 및 패키지 구조 수정
